tou acceptance added to first time flow

+++ app/models/concerns/static_resource.rb
@@ -7,6 +7,7 @@ module StaticResource
 	end
 
 	def clear_all_caches
+		return if Rails.env.development?
 		Rails.cache.delete_matched('*', namespace: self.class.name)
 	end
